Finally fix the Live macros to only do stuff inside the IDE!!!
The big insight was dispatching on a separate _File Type_ for LiveIDE
files in :macrocall, so that the overridden Live macros are only evoked
from within an active file in the IDE. :)

This prevents them from running when you `include` a file, for example.

Currently have `Live.@test` and `Live.@testfile` working!