Performance Improvement: Cancel in-flight callbacks on new keypress
This prevents expensive callbacks from adding up (such as when you hold
enter to move a block down 10 lines).

Correctness: This also prevents infinite loops from freezing your
program (as long as the infinite loop has any yield-points in it),
because if you just delete it, the callback will be cancelled. :)

(For the sake of the above, I may want to consider artificially
injecting yield-points into the user code? Though this wouldn't help if
you call a function that infinite loops, but is defined in another
file... I probably just need a cancel button, to be honest.)